Bucket implementation fixed to work in 32bit
The last commit worked in 64bit arch but not in 32bit.
This fix works only in 32bit, then do not update the binaries
already compiled to 64bit.

+++ fill/eggfill.c
@@ -122,7 +122,7 @@ void queue_dequeue(queue *q){
 }/* end of queue*/
 
 void
-floodfill(unsigned int* pixels, int x, int y, int width, int height, unsigned int color) {
+floodfill(unsigned long * pixels, int x, int y, int width, int height,unsigned long color) {
 
     printf("\nEntrando to floodfill\n");
     queue *lista_xy;
